rename gameEventsBroker module to gameEventsHandler
