+++ src/chat/chat.module.ts
@@ -6,13 +6,13 @@ import {ChatService} from "./chat.service";
 import {isChatExists} from "./decorator/validation/IsChatExists.decorator";
 import {ModelName} from "../common/enum/modelName.enum";
 import {RequestHelperModule} from "../requestHelper/requestHelper.module";
-import { GameEventsBrokerModule } from '../gameEventsBroker/gameEventsBroker.module';
+import { GameEventsHandlerModule } from '../gameEventsHandler/gameEventsHandler.module';
 
 @Module({
     imports: [
         MongooseModule.forFeature([ {name: ModelName.CHAT, schema: ChatSchema} ]),
         RequestHelperModule,
-        GameEventsBrokerModule
+        GameEventsHandlerModule
     ],
     controllers: [ChatController],
     providers: [ ChatService, isChatExists ],
