+++ src/gameEventsHandler/clanEventHandler.ts
